vws_taco_api/requirements.txt
Hug==2.3.0
SQLAlchemy==1.1.13
Flask==0.12.2
PyJWT==1.5.3
argon2_cffi==18.1.0